feat!: accept generic and text-to-text functions in same argument
Accept both types of inference functions, generic and
text-to-text, in the same argument rather than using separate
keyword arguments for each.

A) It's simpler. B) It's more Pythonic to "ask for forgiveness
rather than permission" by hitting an exception when accessing
the wrong type of function rather than testing for the right
function before calling it. And C) This will scale easier: this
single argument can be extended to also accept a heterogeneous
list of inference functions from which multiple library blocks
can be dynamically created.